+++ src/modelling/sensor-simulators/PositionSensorSimulator.cpp
@@ -1,5 +1,5 @@
 #include "PositionSensorSimulator.h"
 
-PositionSensorSimulator::PositionSensorSimulator(std::vector<Vector3f> path, float deltaTimeSeconds) {
+PositionSensorSimulator::PositionSensorSimulator(std::vector<Vector3d> path, double deltaTimeSeconds) {
   m_data = path;
 }
